refacto: create input_api
